Extract promotion and ship CLI
